requirements.txt
Chameleon==2.14
Jinja2==2.7.1
Mako==0.9.1
MarkupSafe==0.18
PasteDeploy==1.5.2
Pillow==2.2.2
PyYAML==3.10
Pygments==1.6
SQLAlchemy==0.8.4
WebOb==1.3.1
WebTest==2.0.14
amqplib==1.0.2
anyjson==0.3.3
beautifulsoup4==4.3.2
caldav==0.1.12
celery==2.5.3
coverage==3.7.1
cryptacular==1.4.1
icalendar==3.5
ipython==1.2.1
jsonschema==0.7
kombu==2.1.8
lxml==3.3.2
nose==1.3.0
pbkdf2==1.3
pyramid==1.4
pyramid-filterwarnings==0.4
pyramid-jinja2==1.9
pyramid-mako==0.3.1
pyramid-tm==0.7
python-dateutil==1.5
python-ldap==2.4.14
pytz==2013.9
pyvac==0.1
redis==2.8.0
repoze.lru==0.6
simplejson==2.6.2
six==1.4.1
transaction==1.4.1
translationstring==1.1
venusian==1.0a8
vobject==0.6.6
waitress==0.8.8
zope.deprecation==4.1.0
zope.interface==4.0.5
zope.sqlalchemy==0.7.3
